requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
streamlit>=1.32.0
pydantic>=2.6.0
numpy>=1.26.0
scikit-learn>=1.4.0
rank-bm25>=0.2.2
pypdf>=4.0.0
pillow>=10.0.0
pytesseract>=0.3.10
requests>=2.31.0

